feat: add LoginWindowViewModel
Added ViewModelBase class, which is the root class in the hierarchy of ViewModels and implements the INotifyPropertyChanged interface. Also implemented events to display the message box, close window.
Added RelayCommand class, which is the root class in the hierarchy of Commands and implements the ICommand interface.
Solved the problem with password box binding. This method does not violate the MVVM pattern and maintains complete security.